feat: include endpoint to allow associate to vote

+++ src/main/resources/db/migration/V3__CREATE_TABLE_VOTING_SESSION.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS public.voting_session (
     duration timestamp(6) NOT NULL,
     description VARCHAR(1000) NULL,
     is_open bool NOT NULL DEFAULT true,
-    voting_topic_id INT,
+    voting_topic_id bigserial,
     created_at timestamp(6) NULL,
     FOREIGN KEY (voting_topic_id) REFERENCES voting_topic(id)
 );
